Code CachedConnections and TcpTransportLayer clean up
