+++ comm/src/main/scala/coop/rchain/comm/CachedConnections.scala
@@ -5,14 +5,16 @@ import cats.effect.Concurrent
 import cats.syntax.applicative._
 import cats.syntax.flatMap._
 import cats.syntax.functor._
-import coop.rchain.catscontrib.ski.kp
-import coop.rchain.comm.transport.{TcpTransportLayer, TransportState}
-import coop.rchain.shared.Cell
-import io.grpc.ManagedChannel
 
+import coop.rchain.catscontrib.ski.kp
+import coop.rchain.shared.Cell
+
+import io.grpc.ManagedChannel
 import scala.language.higherKinds
 
-class CachedConnections[F[_], T](val cell: TcpTransportLayer.TransportCell[F])(
+import monix.execution.Cancelable
+
+class CachedConnections[F[_], T](val cell: Transport.TransportCell[F])(
     clientChannel: PeerNode => F[ManagedChannel]
 )(implicit E: MonadError[F, Throwable]) {
 
@@ -34,4 +36,21 @@ object CachedConnections {
     for {
       connections <- Cell.mvarCell[F, TransportState](TransportState.empty)
     } yield new CachedConnections[F, T](connections)(_)
+}
+
+object Transport {
+  type Connection          = ManagedChannel
+  type Connections         = Map[PeerNode, Connection]
+  type TransportCell[F[_]] = Cell[F, TransportState]
+}
+
+case class TransportState(
+    connections: Transport.Connections = Map.empty,
+    server: Option[Cancelable] = None,
+    clientQueue: Option[Cancelable] = None,
+    shutdown: Boolean = false
+)
+
+object TransportState {
+  def empty: TransportState = TransportState()
 }
